Section out the Queue class to its own file
