+++ envision/tests/test_data_replay.py
@@ -21,7 +21,6 @@
 # THE SOFTWARE.
 
 import multiprocessing
-import multiprocessing.queues
 import tempfile
 from pathlib import Path
 
@@ -34,75 +33,12 @@ from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.scenario import Scenario
 from smarts.core.smarts import SMARTS
 from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
+from envision.utils.queue import Queue
 
 AGENT_ID = "Agent-007"
 NUM_EPISODES = 3
 MAX_STEPS = 100
 TIMESTEP_SEC = 0.1
-
-
-class SharedCounter(object):
-    """A synchronized shared counter.
-    The locking done by multiprocessing.Value ensures that only a single
-    process or thread may read or write the in-memory ctypes object. However,
-    in order to do n += 1, Python performs a read followed by a write, so a
-    second process may read the old value before the new one is written by the
-    first process. The solution is to use a multiprocessing.Lock to guarantee
-    the atomicity of the modifications to Value.
-    This class comes almost entirely from Eli Bendersky's blog:
-    http://eli.thegreenplace.net/2012/01/04/shared-counter-with-pythons-multiprocessing/
-    """
-
-    def __init__(self, n=0):
-        self.count = multiprocessing.Value("i", n)
-
-    def increment(self, n=1):
-        """ Increment the counter by n (default = 1) """
-        with self.count.get_lock():
-            self.count.value += n
-
-    @property
-    def value(self):
-        """ Return the value of the counter """
-        return self.count.value
-
-
-class Queue(multiprocessing.queues.Queue):
-    """A portable implementation of multiprocessing.Queue.
-    Because of multithreading / multiprocessing semantics, Queue.qsize() may
-    raise the NotImplementedError exception on Unix platforms like Mac OS X
-    where sem_getvalue() is not implemented. This subclass addresses this
-    problem by using a synchronized shared counter (initialized to zero) and
-    increasing / decreasing its value every time the put() and get() methods
-    are called, respectively. This not only prevents NotImplementedError from
-    being raised, but also allows us to implement a reliable version of both
-    qsize() and empty().
-    """
-
-    def __init__(self, *args, **kwargs):
-        super(Queue, self).__init__(*args, **kwargs)
-        self.size = SharedCounter(0)
-
-    def put(self, *args, **kwargs):
-        self.size.increment(1)
-        super(Queue, self).put(*args, **kwargs)
-
-    def get(self, *args, **kwargs):
-        self.size.increment(-1)
-        return super(Queue, self).get(*args, **kwargs)
-
-    def qsize(self):
-        """ Reliable implementation of multiprocessing.Queue.qsize() """
-        return self.size.value
-
-    def empty(self):
-        """ Reliable implementation of multiprocessing.Queue.empty() """
-        return not self.qsize()
-
-    def clear(self):
-        """ Remove all elements from the Queue. """
-        while not self.empty():
-            self.get()
 
 
 @pytest.fixture
